Working on commands

+++ oop/uim/oop/commands/interfaces.d
@@ -4,5 +4,5 @@ import uim.oop;
 @safe:
 
 interface ICommand :INamed {
-    int execute(Json[string] options, IConsole console = null);
+    ulong execute(Json[string] options, IConsole console = null);
 }
